Add controller de produto na sacola

+++ src/main/java/com/apiecommerce/apiecomerce/server/controllers/AuthenticationController.java
@@ -57,7 +57,7 @@ public class AuthenticationController {
     }
 
     @PutMapping("/register/{id}")
-    public ResponseEntity registerAdmin(@RequestBody LoginResponseDTO loginResponseDTO, @PathVariable int id) {
+    public ResponseEntity registerAdmin(@RequestBody LoginResponseDTO loginResponseDTO, @PathVariable Long id) {
         Usuario usuario = repository.findById(id).get();
         usuario.setRole(Roles.ADMIN);
         repository.saveAndFlush(usuario);
